Fix test run
