+++ tests/wasm_worker/shared_memory.c
@@ -1,6 +1,7 @@
 #include <emscripten.h>
+#include <stdio.h>
 
 int main()
 {
-	_ReportResult(EM_ASM_INT(return buffer instanceof SharedArrayBuffer), 0);
+	printf("%d\n", EM_ASM_INT(return buffer instanceof SharedArrayBuffer));
 }
